pyside_mascotas/requirements.txt

PySide6==6.8.1
qdarkstyle==3.2.3
SQLAlchemy==2.0.36
Alembic==1.14.1
python-dotenv==1.0.1
PyMySQL==1.1.1
cryptography==44.0.1